backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.5
openai==1.78.1
pip==25.0
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pyjwt==2.10.1
python-dotenv==1.1.0
pyyaml==6.0.2
rich==14.0.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
tree-sitter==0.20.1
tree-sitter-javascript==0.23.1
tree-sitter-languages==1.10.2
tree-sitter-python==0.23.6
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
